ARCH-3: Renamed files with extensions, enabled SwiftLint rule 'file_name' and disabled it only in MockedData.swift
