+++ WineUp/WineUp/Models/MockedData.swift
@@ -1,3 +1,4 @@
+//swiftlint:disable:this file_name
 //
 //  MockedData.swift
 //  WineUp
