--- WineUp/WineUp/Modules/ApplicationMenu/ApplicationMenuViewModel.swift
@@ -1,16 +1,0 @@
-//
-//  ApplicationMenuViewModel.swift
-//  WineUp
-//
-//  Created by Александр Пахомов on 24.09.2020.
-//
-
-import Foundation
-
-extension ApplicationMenuView {
-    class ViewModel: ObservableObject {
-        var catalogRootViewModel: CatalogRootView.ViewModel {
-            return .init()
-        }
-    }
-}
