+++ WineUp/WineUp/Modules/ApplicationRoot/ApplicationRootView+ViewModel.swift
@@ -1,5 +1,5 @@
 //
-//  ApplicationRootViewModel.swift
+//  ApplicationRootView+ViewModel.swift
 //  WineUp
 //
 //  Created by Александр Пахомов on 22.09.2020.
