+++ WineUp/WineUp/Modules/Catalog/Catalog/CatalogView+ViewModel.swift
@@ -1,5 +1,5 @@
 //
-//  CatalogViewModel.swift
+//  CatalogView+ViewModel.swift
 //  WineUp
 //
 //  Created by Александр Пахомов on 24.09.2020.
